// src/types/translationTypes.d.ts
type register = {


    teacher: {
        title: string;
        privacy: string;
        button: string,
        inputs: string[];
    },

    student: {

        title: string,
        subtitle: string,
        privacy: string,
        button: string,
        aside: {
            title: string,
            items: Array<string>
        }
        form: {
            name: string,
            lastName: string,
            email: string,
            nativeLanguage: string,
            countryCode: string,
            phone: string,
            languageToLearn: string,
            submit: string
        }

    },
    login: {
        title: string,
        registerTeacher: string,
        registerStudent: string,
        or: string
        form: {
            email: string,
            password: string,
            forgot: string,
            remember: string
        }
        button: string
    },
    recover: {
        email: string,
        button: string
    }

}